+++ packages/blocks-core/js/libs/wordpress/avatar/test/functionality.blocks-1.e2e.cy.js
@@ -142,8 +142,12 @@ describe('Avatar Block', () => {
 		cy.getByDataTest('popover-body')
 			.first()
 			.within(() => {
-				cy.get('input[maxlength="9"]').clear({ force: true });
-				cy.get('input[maxlength="9"]').type('37e6d4', { delay: 0 });
+				cy.get('[data-cy="color-picker-css-value"]').clear({
+					force: true,
+				});
+				cy.get('[data-cy="color-picker-css-value"]').type('37e6d4', {
+					delay: 0,
+				});
 			});
 
 		cy.get('@borderContainer').within(() => {