+++ packages/controls/js/libs/filter-control/test/filter-control.cy.js
@@ -179,8 +179,12 @@ describe('filter-control component testing', () => {
 
 			cy.getByDataTest('filter-drop-shadow-color').click();
 			cy.contains('Color Picker').as('color-picker');
-			cy.get('@color-picker').get('input[maxlength="9"]').clear();
-			cy.get('@color-picker').get('input[maxlength="9"]').type('2cf1dd');
+			cy.get('@color-picker')
+				.find('[data-cy="color-picker-css-value"]')
+				.clear();
+			cy.get('@color-picker')
+				.find('[data-cy="color-picker-css-value"]')
+				.type('2cf1dd');
 
 			//Check value
 			cy.get('select').eq(0).should('have.value', 'drop-shadow');