+++ packages/dev-cypress/js/support/commands.js
@@ -566,8 +566,10 @@ export const registerCommands = () => {
 		cy.get('[data-wp-component="Popover"]')
 			.last()
 			.within(() => {
-				cy.get('input[maxlength="9"]').clear({ force: true });
-				cy.get('input[maxlength="9"]')
+				cy.get('[data-cy="color-picker-css-value"]').clear({
+					force: true,
+				});
+				cy.get('[data-cy="color-picker-css-value"]')
 					.type(value + ' ', { delay: 0 })
 					.then(() => {
 						if (Cypress.$(`[aria-label="Close"]`).length) {