+++ packages/editor/js/extensions/libs/block-card/block-states/test/block-states.general-2.e2e.cy.js
@@ -485,10 +485,13 @@ describe('Block State E2E Test', () => {
 			cy.getByDataTest('popover-body')
 				.last()
 				.within(() => {
-					cy.get('input[maxlength="9"]').clear({
+					cy.get('[data-cy="color-picker-css-value"]').clear({
 						force: true,
 					});
-					cy.get('input[maxlength="9"]').type('000000', { delay: 0 });
+					cy.get('[data-cy="color-picker-css-value"]').type(
+						'000000',
+						{ delay: 0 }
+					);
 				});
 
 			context(
@@ -500,12 +503,15 @@ describe('Block State E2E Test', () => {
 					cy.getByDataTest('popover-body')
 						.last()
 						.within(() => {
-							cy.get('input[maxlength="9"]').clear({
+							cy.get('[data-cy="color-picker-css-value"]').clear({
 								force: true,
 							});
-							cy.get('input[maxlength="9"]').type('cccccc', {
-								delay: 0,
-							});
+							cy.get('[data-cy="color-picker-css-value"]').type(
+								'cccccc',
+								{
+									delay: 0,
+								}
+							);
 						});
 
 					// inherit of normal.